+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- パフォーマンス検索条件から作品関係を削除
+- パフォーマンスコレクションからリレーションを削除
+- 予約リポジトリのmongooseモデルをprivate化
+
 ### Deprecated
 
 ### Removed